+++ src/bin/inx-chronicle/stardust_inx/mod.rs
@@ -35,6 +35,8 @@ pub struct InxWorker {
     db: MongoDb,
     #[cfg(any(feature = "analytics", feature = "metrics"))]
     influx_db: Option<chronicle::db::influxdb::InfluxDb>,
+    #[cfg(any(feature = "analytics", feature = "metrics"))]
+    influx_db_config: Option<chronicle::db::influxdb::InfluxDbConfig>,
     config: InxConfig,
 }
 
@@ -85,12 +87,15 @@ impl InxWorker {
     pub fn new(
         db: &MongoDb,
         #[cfg(any(feature = "analytics", feature = "metrics"))] influx_db: Option<&chronicle::db::influxdb::InfluxDb>,
+        #[cfg(any(feature = "analytics", feature = "metrics"))] influx_db_config: Option<&chronicle::db::influxdb::InfluxDbConfig>,
         inx_config: &InxConfig,
     ) -> Self {
         Self {
             db: db.clone(),
             #[cfg(any(feature = "analytics", feature = "metrics"))]
             influx_db: influx_db.cloned(),
+            #[cfg(any(feature = "analytics", feature = "metrics"))]
+            influx_db_config: influx_db_config.cloned(),
             config: inx_config.clone(),
         }
     }
@@ -112,8 +117,16 @@ impl InxWorker {
 
         debug!("Started listening to ledger updates via INX.");
 
-        #[cfg(feature = "analytics")]
-        let mut analytics = chronicle::db::collections::analytics::all_analytics();
+        #[cfg(any(feature = "analytics", feature = "metrics"))]
+        let mut selected_analytics = if self.influx_db_config.as_ref().unwrap().selected_analytics.is_empty() {
+            chronicle::db::collections::analytics::all_analytics()
+        } else {
+            tracing::info!("Computing the following analytics: {:?}", self.influx_db_config.as_ref().unwrap().selected_analytics);
+
+            let mut tmp: std::collections::HashSet<chronicle::db::influxdb::config::AnalyticsChoice> =
+                self.influx_db_config.as_ref().unwrap().selected_analytics.iter().copied().collect();
+            tmp.drain().map(Into::into).collect()
+        };
 
         while let Some(ledger_update) = stream.try_next().await? {
             self.handle_ledger_update(
@@ -121,7 +134,7 @@ impl InxWorker {
                 ledger_update,
                 &mut stream,
                 #[cfg(feature = "analytics")]
-                &mut analytics,
+                &mut selected_analytics,
             )
             .await?;
         }
@@ -274,7 +287,7 @@ impl InxWorker {
         inx: &mut Inx,
         start_marker: LedgerUpdateMessage,
         stream: &mut (impl futures::Stream<Item = Result<LedgerUpdateMessage, InxError>> + Unpin),
-        #[cfg(feature = "analytics")] analytics: &mut Vec<Box<dyn chronicle::db::collections::analytics::Analytic>>,
+        #[cfg(feature = "analytics")] selected_analytics: &mut Vec<Box<dyn chronicle::db::collections::analytics::Analytic>>,
     ) -> Result<()> {
         #[cfg(feature = "metrics")]
         let start_time = std::time::Instant::now();
@@ -386,7 +399,7 @@ impl InxWorker {
         #[cfg(feature = "analytics")]
         if let Some(influx_db) = &self.influx_db {
             if influx_db.config().analytics_enabled {
-                gather_analytics(&self.db, influx_db, analytics, milestone_index, milestone_timestamp).await?;
+                gather_analytics(&self.db, influx_db, selected_analytics, milestone_index, milestone_timestamp).await?;
             }
         }
         #[cfg(all(feature = "analytics", feature = "metrics"))]
