+++ src/db/influxdb/mod.rs
@@ -9,7 +9,7 @@ use std::ops::Deref;
 use influxdb::{Client, ReadQuery};
 use serde::de::DeserializeOwned;
 
-pub use self::{config::InfluxDbConfig, measurement::InfluxDbMeasurement};
+pub use self::{config::{InfluxDbConfig, AnalyticsChoice}, measurement::InfluxDbMeasurement};
 
 /// A wrapper for an InfluxDb [`Client`].
 #[derive(Clone, Debug)]
